moved all possible logic to the base class from Direct and Debye classes
